Add files via upload
The project uses the OpenCV library in python to detect face from an image or a video. We are using haarcascades to detect various features of a human body like the eyes,face , nose etc.
Please make sure that you download all the files and put them in the same folder in which you have your project, otherwise the code will not detect these files and it will not run.
Ensure that opencv library is installed in your IDE (Anaconda, pycharm etc). Without installing this library you cannot import the "cv2" file and many image processing functions will not work thereby causing various errors